Revert "types"
This reverts commit b9f9723c8f15f7622addd7c8a8f0692b8fcfeb44.

+++ include/seqan/kmer/kmer_ibf.h
@@ -222,7 +222,7 @@ public:
      * \param counts Vector to be filled with counts.
      * \param text Text to count occurences for.
      */
-    void select(std::vector<uint32_t> & counts, TString const & text) // TODO uint16_t
+    void select(std::vector<uint16_t> & counts, TString const & text) // TODO uint16_t
     {
         uint32_t possible = length(text) - kmerSize + 1; // Supports text lengths up to 65535 + k
